requirements.txt
# Core agent
langchain>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0

# Browser automation
playwright>=1.44.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9   # for PostgreSQL (optional, SQLite works by default)

# LLM / AI
groq>=0.9.0

# PDF generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0